embedding/image/requirements.txt
flatbuffers==25.12.19
mpmath==1.3.0
packaging==26.0
protobuf==6.33.5
sympy==1.14.0
numpy==1.26.4
onnxruntime==1.16.3
pillow==10.2.0
supabase